detail cargado del server
