+++ client/src/app/Appraisals/page.jsx
@@ -3,6 +3,7 @@ import { useState } from 'react';
 
 import { validateForm } from '../Appraisals/validates';
 import axios from 'axios';
+import Title from "./title"
 
 export default function Example() {
   const [formData, setFormData] = useState({
@@ -54,19 +55,7 @@ export default function Example() {
   return (
     <div class=" min-h-screen flex items-center justify-center bg-center py-12 px-4 sm:px-6 lg:px-8 bg-black bg-no-repeat bg-cover ">
       <div className=" mx-auto md:max-w-35rem h-50 space-y-8 p-10 bg-white rounded-xl shadow-lg z-10">
-        <div className="mx-auto max-w-2xl lg:text-center">
-          <h2 className="text-base font-semibold leading-7 text-indigo-600">
-            Nosotros
-          </h2>
-          <p className="mt-2 text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl">
-            Formulario Tasaciones
-          </p>
-          <p className="mt-6 text-lg leading-8 text-gray-600">
-            Puedes dejar tu peticion para que hagamos una tasacion de tu
-            propiedad, a la brevedad nos contactaremos contigo para ultimar
-            detalles.
-          </p>
-        </div>
+        <Titulo/>
         <div className="mx-auto mt-16 max-w-2xl sm:mt-20 lg:mt-24 lg:max-w-4xl">
           <dl className=" max-w-xl mr-12 gap-x-8 gap-y-10 lg:max-w-none lg:grid-cols-2 lg:gap-y-16">
             <div className="relative pl-16">
