+++ client/src/app/Components/NavBar/Navbar.jsx
@@ -1,4 +1,4 @@
-"use client";
+
 import { Fragment } from "react";
 import { Disclosure, Menu, Transition } from "@headlessui/react";
 import { Bars3Icon, BellIcon, XMarkIcon } from "@heroicons/react/24/outline";
