+++ client/src/app/Detail/[id]/page.jsx
@@ -1,30 +1,38 @@
-"use client";
-import { useState, useEffect } from "react";
+
+// import { useState, useEffect } from "react";
 import axios from "axios";
 import styles from "./page.module.css";
-import { GoogleMap, LoadScript, Marker } from "@react-google-maps/api";
 
-export default function Example({ params }) {
-  const [propiedad, setPropiedad] = useState();
-  const [modalOpen, setModalOpen] = useState(false);
-  const [currentIndex, setCurrentIndex] = useState(0);
-  const [markerPosition, setMarkerPosition] = useState(null);
+import Gallery from "./gallery"
+import { idText } from "typescript";
+import Map from "./map"
 
+
+
+const fetchPropId = async  (id) =>{
+  const res = await fetch(`http://localhost:3001/realState/${id}`)
+         return  res.json()
+         
+        }
+
+export default async function Example({ params }) {
   const { id } = params;
+ const propiedad = await fetchPropId(id)
+ 
 
-  useEffect(() => {
-    axios.get(`http://localhost:3001/realState/${id}`).then((res) => {
-      if (res.data) {
-        setPropiedad(res.data);
-      } else {
-        window.alert("No hay propiedades con ese ID");
-      }
-    });
-  }, [id]);
 
-  if (!propiedad) {
-    return <div>Cargando...</div>;
-  }
+
+
+
+
+  
+  console.log(propiedad);
+ 
+
+
+
+
+
 
   const handleReservarClick = async () => {
     try {
@@ -43,10 +51,10 @@ export default function Example({ params }) {
 
   const product = {
     name: propiedad.address,
-    price: "$ " + propiedad.operations[0].prices[0].price,
+    price: "$ " + propiedad.price,
     href: "#",
     breadcrumbs: [
-      { id: 1, name: propiedad.operations[0].operation_type },
+      { id: 1, name: propiedad.operation_type },
       { id: 2, name: propiedad.type },
     ],
     description: propiedad.description,
@@ -61,39 +69,22 @@ export default function Example({ params }) {
     propiedad.photos[5],
   ];
 
-  const openModal = (index) => {
-    setCurrentIndex(index);
-    setModalOpen(true);
-  };
 
-  const closeModal = () => {
-    setModalOpen(false);
-  };
+  
+const center = {
+  lat: parseFloat(propiedad?.geo_lat),
+  lng: parseFloat(propiedad?.geo_long),
+};
 
-  const changeImage = (n) => {
-    let newIndex = currentIndex + n;
-    if (newIndex < 0) {
-      newIndex = images.length - 1;
-    } else if (newIndex >= images.length) {
-      newIndex = 0;
-    }
-    setCurrentIndex(newIndex);
-  };
 
-  const mapContainerStyle = {
+
+   const mapContainerStyle = {
     marginTop: "50px",
     width: "100%",
     height: "400px",
   };
 
-  const center = {
-    lat: parseFloat(propiedad.geo_lat),
-    lng: parseFloat(propiedad.geo_long),
-  };
 
-  const handleMapLoad = (map) => {
-    setMarkerPosition(center);
-  };
 
   return (
     <div className="bg-white">
@@ -138,41 +129,11 @@ export default function Example({ params }) {
         </nav>
 
         {/* Image gallery */}
-        <div className="container mx-auto px-2 py-2 lg:px-32 lg:pt-12">
-          <div className="flex flex-wrap md:-m-2">
-            {images.map((image, index) => (
-              <div className="w-1/3 p-1 md:p-2" key={index}>
-                <img
-                  alt="gallery"
-                  className="block h-full w-full rounded-lg object-cover object-center cursor-pointer"
-                  src={image}
-                  onClick={() => openModal(index)}
-                />
-              </div>
-            ))}
-          </div>
+        <Gallery images= {images}/>
+        
 
-          {modalOpen && (
-            <div className={styles.modalOverlay}>
-              <div className={styles.modal}>
-                <span className={styles.close} onClick={closeModal}>
-                  &times;
-                </span>
-                <img
-                  className={styles.modalContent}
-                  src={images[currentIndex]}
-                  alt="modal"
-                />
-                <a className={styles.prev} onClick={() => changeImage(-1)}>
-                  &#10094;
-                </a>
-                <a className={styles.next} onClick={() => changeImage(1)}>
-                  &#10095;
-                </a>
-              </div>
-            </div>
-          )}
-        </div>
+
+
 
         {/* Product info */}
         <div className="mx-auto max-w-2xl px-4 pb-16 pt-10 sm:px-6 lg:grid lg:max-w-7xl lg:grid-cols-3 lg:grid-rows-[auto,auto,1fr] lg:gap-x-8 lg:px-8 lg:pb-24 lg:pt-16">
@@ -244,13 +205,13 @@ export default function Example({ params }) {
               </div>
             </div>
 
-            <button
+            {/* <button
               type="submit"
               className="mt-10 flex w-full items-center justify-center rounded-md border border-transparent bg-indigo-600 px-8 py-3 text-base font-medium text-white hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
               onClick={handleReservarClick}
             >
               Reservar
-            </button>
+            </button> */}
           </div>
 
           <div className="py-10 lg:col-span-2 lg:col-start-1 lg:border-r lg:border-gray-200 lg:pb-16 lg:pr-8 lg:pt-6">
@@ -262,7 +223,8 @@ export default function Example({ params }) {
                 <p className="text-base text-gray-900">{product.description}</p>
               </div>
             </div>
-            <LoadScript googleMapsApiKey="AIzaSyDSHb1gBxjiShwN6LVC_bBnxt9E60TwvmI">
+            <Map mapContainerStyle = {mapContainerStyle} center = {center}/>
+            {/* <LoadScript googleMapsApiKey="AIzaSyDSHb1gBxjiShwN6LVC_bBnxt9E60TwvmI">
               <GoogleMap
                 mapContainerStyle={mapContainerStyle}
                 center={center}
@@ -271,7 +233,7 @@ export default function Example({ params }) {
               >
                 <Marker position={markerPosition} />
               </GoogleMap>
-            </LoadScript>
+            </LoadScript> */}
           </div>
         </div>
       </div>
